Change the way we run commands under NVM
I've started invoking commands that need to run under NVM/RVM by
prefixing them with 'bash -lc' so they're run under a login shell, which
picks up all the NVM/RVM shell magic correctly. This change updates this
role to do this too.

Additionally it changes the default node version detection so it's not
always a failure (that's ignored). I did this so it was easier to work
with the above invocation magic but I also think it's cleaner and a more
Ansible way to do things.